Refactor api/contributor to use GH proxy
